Move PA types to common package.
